pyproject.toml
[project]
name = "olake-slack-agent"
version = "1.0.0"
description = "AI-powered Slack community agent for OLake support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "slack-sdk>=3.27.0",
    "flask>=3.0.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "google-generativeai>=0.3.0",
    "openai>=1.10.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
olake-agent = "agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
